move EFile class to its own file
